Use a thread for TAP reads that uses an event to trigger the main event loop.
A custom version of handle_device_data() is used that resets the event and
consumes any waiting packets.  There is no need to pass the number of wait
packets, we simply consume as many as are available.

This is a rather simple implementation, we can probably get better
performance by using multiple simultaneous reads.